Based functionality for dynamic roles setting Co-authored-by: Bao Pham <BaoPham92@yahoo.com>
