--- .env.sample
@@ -1,2 +1,0 @@
-ST_DATABASE_URL=<add_sauti_database_url>
-MOESIF_ID=<add_moesif_id>
