+++ routes/apikeyRoute.js
@@ -2,12 +2,11 @@ const express = require('express')
 const router = express.Router()
 const uuidAPIKey = require('uuid-apikey')
 const bcrypt = require('bcryptjs')
-
 const db = require('../api-key/dbConfig')
-
 const jwtCheck = require('../middleware/token-middleware')
+const rules = require('../middleware/rules/rules-middleware');
 
-router.post('/private', jwtCheck, async (req, res) => {
+router.post('/private', jwtCheck, rules, async (req, res) => {
   const key = uuidAPIKey.create()
   const { id } = req.body
